Update: DELLETE /api/v1/follow API exception handling 추가
자신의 FOLLOWEE, FOLLOWING 정보가 아니라면 FOLLOW_ACCESS_UNAUTHORIZED를 던지도록 추가함.

+++ src/main/java/com/divide/follow/FollowService.java
@@ -33,9 +33,13 @@ public class FollowService {
         return newFollow.getId();
     }
 
-    public Long remove(Long followId) {
+    public Long remove(Long followId, User me) throws RestApiException {
         Follow follow = followRepository.findById(followId)
                 .orElseThrow(() -> new RestApiException(FollowErrorCode.FOLLOW_NOT_FOUND));
+
+        if (!follow.getFollower().equals(me) && !follow.getFollowee().equals(me)) {
+            throw new RestApiException(FollowErrorCode.FOLLOW_ACCESS_UNAUTHORIZED);
+        }
 
         followRepository.remove(follow);
         return followId;
